googlesheets4: Access Google Sheets using the Sheets API V4

Interact with Google Sheets through the Sheets API v4 https://developers.google.com/sheets/api. "API" is an acronym for "application programming interface"; the Sheets API allows users to interact with Google Sheets programmatically, instead of via a web browser. The "v4" refers to the fact that the Sheets API is currently at version 4. This package can read and write both the metadata and the cell data in a Sheet.

See also

Author

Maintainer: Jennifer Bryan jenny@posit.co (ORCID)

Other contributors:

Posit Software, PBC [copyright holder, funder]