Class for Google Sheets formulas

In order to write a formula into Google Sheets, you need to store it as an object of class googlesheets4_formula. This is how we distinguish a "regular" character string from a string that should be interpreted as a formula. googlesheets4_formula is an S3 class implemented using the vctrs package.

gs4_formula(x = character())

Arguments

x
Character.

Value

An S3 vector of class googlesheets4_formula.

See also

Examples

if (gs4_has_token()) {
  dat <- data.frame(X = c(1, 5, 3, 2, 4, 6))

  ss <- sheet_write(dat)
  ss

  summaries <- tibble::tribble(
    ~desc, ~summaries,
    "max", "=max(A:A)",
    "sum", "=sum(A:A)",
    "min", "=min(A:A)",
    "sparkline", "=SPARKLINE(A:A, {\"color\", \"blue\"})"
  )

  # explicitly declare a column as `googlesheets4_formula`
  summaries$summaries <- gs4_formula(summaries$summaries)
  summaries

  range_write(ss, data = summaries, range = "C1", reformat = FALSE)

  miscellany <- tibble::tribble(
    ~desc, ~example,
    "hyperlink", "=HYPERLINK(\"http://www.google.com/\",\"Google\")",
    "image", "=IMAGE(\"https://www.google.com/images/srpr/logo3w.png\")"
  )
  miscellany$example <- gs4_formula(miscellany$example)
  miscellany

  sheet_write(miscellany, ss = ss)

  # clean up
  googledrive::drive_trash(ss)
}

#> Creating new Sheet: "unfearing-guineafowl"

#> Editing "unfearing-guineafowl"

#> Writing to sheet "dat"

#> Changing dims: (7 x 1) --> (7 x 4)

#> Writing to "unfearing-guineafowl"

#> Writing to sheet "miscellany"

#> Files trashed:
#>   * unfearing-guineafowl: 1cP6J-5nlfEr8URqyRBDtIr-898nOdvqFXgl3EHc_gPQ