gs4_get {googlesheets4}	R Documentation

Get Sheet metadata

Description

Retrieve spreadsheet-specific metadata, such as details on the individual (work)sheets or named ranges.

gs4_get() complements googledrive::drive_get(), which returns metadata that exists for any file on Drive.

Usage

gs4_get(ss)

Arguments

ss
Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

Value

A list with S3 class googlesheets4_spreadsheet, for printing purposes.

See Also

Wraps the spreadsheets.get endpoint:

https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/get

Examples

gs4_get(gs4_example("mini-gap"))

[Package googlesheets4 version 1.1.1 Index]