Produce configured token

For internal use or for those programming around the Sheets API. Returns a token pre-processed with httr::config(). Most users do not need to handle tokens "by hand" or, even if they need some control, gs4_auth() is what they need. If there is no current token, gs4_auth() is called to either load from cache or initiate OAuth2.0 flow. If auth has been deactivated via gs4_deauth(), gs4_token() returns NULL.

Usage

gs4_token()

Value

A request object (an S3 class provided by httr).

See also

Examples

req <- request_generate(
  "sheets.spreadsheets.get",
  list(spreadsheetId = "abc"),
  token = gs4_token()
)
req
#> $method
#> [1] "GET"
#> 
#> $url
#> [1] "https://sheets.googleapis.com/v4/spreadsheets/abc"
#> 
#> $body
#> named list()
#> 
#> $token
#> <request>
#> Auth token: TokenServiceAccount
#>