Get data about (work)sheets

Reveals full metadata or just the names for the (work)sheets inside a (spread)Sheet.

Usage,

sheet_properties(ss)

sheet_names(ss)

Arguments

ss

Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

Value

sheet_properties(): A tibble with one row per (work)sheet.

sheet_names(): A character vector of (work)sheet names.

See also

Examples

if (gs4_has_token()) {
  ss <- gs4_example("gapminder")
  sheet_properties(ss)
  sheet_names(ss)
}
#> [1] "Africa"   "Americas" "Asia"     "Europe"   "Oceania"