View or edit auth config

These functions give the user more control over auth than what is possible with sheets_auth(). Learn more in Google's documentation: Credentials, access, security, and identity and Using OAuth 2.0 for Installed Applications. sheets_auth_config() gives control of:

The OAuth app. If you want to use your own app, setup a new project in Google Developers Console. Follow the instructions in OAuth 2.0 for Mobile & Desktop Apps to obtain your own client ID and secret. Either make an app from your client ID and secret via httr::oauth_app() or provide a path to the JSON file containing same, which you can download from Google Developers Console.

The API key. If googlesheets4 is deauthorized via sheets_deauth(), all requests will be sent with an API key in lieu of a token. If you want to provide your own API key, setup a project as described above and follow the instructions in Setting up API keys.

sheets_api_key() and sheets_oauth_app() retrieve the currently configured API key and OAuth app, respectively.

sheets_auth_config(app = NULL, path = NULL, api_key = NULL)

sheets_api_key()

sheets_oauth_app()

Arguments

app
OAuth app. Defaults to a tidyverse app.

path
Path to the JSON file.

api_key
API key. Defaults to a tidyverse key. Necessary in order to make unauthorized "token-free" requests for public resources.

Value

sheets_auth_config(): An object of R6 class AuthState, which is defined in the gargle package. sheets_api_key(): the current API key. sheets_oauth_app(): the current httr::oauth_app().

See also

Examples

## this will print current config
sheets_auth_config()

#> <AuthState (via gargle)>
#>          <package> googlesheets4
#>              <app> 
#>          <api_key> 
#>      <auth_active> FALSE
#>      <credentials> NULL
#> ---

if (require(httr)) {
  ## bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  sheets_auth_config(app = google_app)
}

#> Loading required package: httr

#> <AuthState (via gargle)>
#>          <package> googlesheets4
#>              <app> my-awesome-google-api-wrapping-package
#>          <api_key> 
#>      <auth_active> FALSE
#>      <credentials> NULL
#> ---

# NOT RUN {
## bring your own app via JSON downloaded from Google Developers Console
sheets_auth_config(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)

sheets_auth_config(api_key = "123")
sheets_api_key()
# }