Edit and view auth configuration

These functions give more control over and visibility into the auth configuration than sheets_auth() does. sheets_auth_configure() lets the user specify their own:

OAuth app, which is used when obtaining a user token.

API key. If googlesheets4 is de-authorized via sheets_deauth(), all requests are sent with an API key in lieu of a token. See the vignette How to get your own API credentials for more. If the user does not configure these settings, internal defaults are used. sheets_oauth_app() and sheets_api_key() retrieve the currently configured OAuth app and API key, respectively.

sheets_auth_configure(app, path, api_key)

sheets_api_key()

sheets_oauth_app()

Arguments

app
OAuth app, in the sense of httr::oauth_app().

path
JSON downloaded from Google Cloud Platform Console, containing a client id (aka key) and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

api_key
API key.

Value

sheets_auth_configure(): An object of R6 class gargle::AuthState, invisibly.

sheets_oauth_app(): the current user-configured httr::oauth_app().

sheets_api_key(): the current user-configured API key.

See also

Examples

# see and store the current user-configured OAuth app (probaby `NULL`)
(original_app <- sheets_oauth_app())

#> NULL

# see and store the current user-configured API key (probaby `NULL`)
(original_api_key <- sheets_api_key())

#> NULL

if (require(httr)) {
  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  google_key <- "the-key-I-got-for-a-google-API"
  sheets_auth_configure(app = google_app, api_key = google_key)

  # confirm the changes
  sheets_oauth_app()
  sheets_api_key()
}

#> Loading required package: httr

#> [1] "the-key-I-got-for-a-google-API"

if (FALSE) {
## bring your own app via JSON downloaded from Google Developers Console
sheets_auth_configure(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)
}

# restore original auth config
sheets_auth_configure(app = original_app, api_key = original_api_key)