Auto-resize columns or rows

Applies automatic resizing to either columns or rows of a (work)sheet. The width or height of targeted columns or rows, respectively, is determined from the current cell contents. This only affects the appearance of a Sheet in the browser and doesn't affect its values in any way.

sheets_auto_resize_dims(
  ss,
  sheet = NULL,
  range = NULL,
  dimension = c("columns", "rows")
)

Arguments

ss
Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get(), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().

sheet
Sheet to modify, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to the first visible sheet.

range
Which columns or rows to resize. Optional. If you want to resize all columns or all rows, use dimension instead. All the usual range specifications are accepted, but the targeted range must specify only columns (e.g. "B:F") or only rows (e.g. "2:7").

dimension
Ignored if range is given. If consulted, dimension must be either "columns" (the default) or "rows". This is the simplest way to request auto-resize for all columns or all rows.

Value

The input ss, as an instance of sheets_id

See also

Examples

if (sheets_has_token()) {
dat <- tibble::tibble(
  fruit = c("date", "lime", "pear", "plum")
)

ss <- sheets_write(dat)
ss

# open in the browser
sheets_browse(ss)

# shrink column A to fit the short fruit names
sheets_auto_resize_dims(ss)

# send some longer fruit names
dat2 <- tibble::tibble(
  fruit = c("cucumber", "honeydew")
)
sheets_append(dat2, ss)
# in the browser, see that column A is now too narrow to show the data

sheets_auto_resize_dims(ss)
# in the browser, see the column A reveals all the data now

googledrive::drive_trash(ss)
}

#> Creating new Sheet: "unfearing-guineafowl"

#> Editing "unfearing-guineafowl"

#> Resizing one or more columns in "dat"

#> Writing to "unfearing-guineafowl"

#> Appending 2 row(s) to "dat"

#> Editing "unfearing-guineafowl"

#> Resizing one or more columns in "dat"

#> Files trashed:
#>   * unfearing-guineafowl: 1nARkmRo27hmkIOYvwDrO4k7ETALrVv00_5rD-5kbxbI