Create a new Sheet

Creates an entirely new (spread)Sheet (or, in Excel-speak, workbook). Offers some control over the initial set of (work)sheets (a.k.a. "tabs").

sheets_create(name, ..., sheets = NULL)

Arguments

name
The name of the new spreadsheet.

...
Optional spreadsheet properties that can be set through this API endpoint, such as locale and time zone.

sheets
Optional named list of data frames. One sheet is created for each data frame.

Value

The input ss, as an instance of sheets_id

See also

Examples

if (sheets_has_token()) {
  sheets_create("sheets-create-demo-1")

  sheets_create("sheets-create-demo-2", locale = "en_CA")

  sheets_create(
    "sheets-create-demo-3",
    locale = "fr_FR",
    timeZone = "Europe/Paris"
  )

  sheets_create(
    "sheets-create-demo-4",
    sheets = list(iris = head(iris), mtcars = head(mtcars))
  )

  # clean up
  sheets_find("sheets-create-demo") %>% googledrive::drive_trash()
}

#> Files trashed:
#>   * sheets-create-demo-4: 1s1q4edMPXBovPcS0q5HSurpSC5mD83TR764d_mCn2bI
#>   * sheets-create-demo-3: 1N20sSsaENKmLu32NBUpWa9aXZxnDgDmtqQawlmSegZE
#>   * sheets-create-demo-2: 1p0_Wx-pKQRnytGQb6D4IicKR4a0UPXedIvUj7k4MDZ8
#>   * sheets-create-demo-1: 1ZqkOPz0TXZCghFfqqQ2-Evx_2lOM0hWVHadyOXEZA8w