Delete cells

Deletes a range of cells and shifts other cells into the deleted area. There are several related tasks that are done by other functions:

To clear cells of their value and/or format, use sheets_clear().

To delete an entire (work)sheet, use sheets_sheet_delete().

To change the dimensions of a (work)sheet, use sheets_sheet_resize().

sheets_delete(ss, sheet = NULL, range, shift = NULL)

Arguments

ss
Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get(), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().

sheet
Sheet to delete, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to the first visible sheet.

range
Cells to delete. There are a couple differences between range here and how it works in other functions (e.g. sheets_read()):

range must be specified.

range must not be a named range.

range must not be the name of a (work) sheet. Use sheets_sheet_delete() to delete an entire sheet. Row-only and column-only ranges are especially relevant, such as "2:6" or "D". Remember you can also use the helpers in cell-specification, such as cell_cols(4:6), or cell_rows(5).

shift
Must be one of "up" or "left", if specified. Required if range is NOT a rows-only or column-only range (in which case, we can figure it out for you). Determines whether the deleted area is filled by shifting surrounding cells up or to the left.

Value

The input ss, as an instance of sheets_id

See also

Examples

if (sheets_has_token()) {
  # create a data frame to use as initial data
  df <- sheets_fodder(10)

  # create Sheet
  ss <- sheets_create("sheets-delete-example", sheets = list(df))

  # delete some rows
  sheets_delete(ss, range = "2:4")

  # delete a column
  sheets_delete(ss, range = "C")

  # delete a rectangle and specify how to shift remaining cells
  sheets_delete(ss, range = "B3:F4", shift = "left")

  # clean up
  googledrive::drive_trash(ss)
}

#> Creating new Sheet: "sheets-delete-example"

#> Editing "sheets-delete-example"

#> Deleting cells in sheet "Sheet1"

#> Editing "sheets-delete-example"

#> Deleting cells in sheet "Sheet1"

#> Editing "sheets-delete-example"

#> Deleting cells in sheet "Sheet1"

#> Files trashed:
#>   * sheets-delete-example: 1zRzCOsbk5UJX5okM_dKw8eeP20Off9aD_Xos4b02ljo