List Sheets endpoints
Source: R/sheets_endpoints.R

Returns a list of selected Sheets API v4 endpoints, as stored inside the googlesheets4 package. The names of this list (or the id sub-elements) are the nicknames that can be used to specify an endpoint in request_generate(). For each endpoint, we store its nickname or id, the associated HTTP method, the path, and details about the parameters. This list is derived programmatically from the Sheets API v4 Discovery Document.

sheets_endpoints(i = NULL)

Arguments

i
The name(s) or integer index(ices) of the endpoints to return. Optional. By default, the entire list is returned.

Value

A list containing some or all of the subset of the Sheets API v4 endpoints that are used internally by googlesheets4.

Examples

str(sheets_endpoints(), max.level = 2)

#> List of 14
#>  $ spreadsheets.get                           :List of 6
#>   ..$ id        : chr "spreadsheets.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}"
#>   ..$ parameters:List of 16
#>   ..$ request   : chr NA
#>   ..$ response  : chr "Spreadsheet"
#>  $ spreadsheets.getByDataFilter               :List of 6
#>   ..$ id        : chr "spreadsheets.getByDataFilter"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}:getByDataFilter"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "GetSpreadsheetByDataFilterRequest"
#>   ..$ response  : chr "Spreadsheet"
#>  $ spreadsheets.create                        :List of 6
#>   ..$ id        : chr "spreadsheets.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets"
#>   ..$ parameters:List of 13
#>   ..$ request   : chr "Spreadsheet"
#>   ..$ response  : chr "Spreadsheet"
#>  $ spreadsheets.batchUpdate                   :List of 6
#>   ..$ id        : chr "spreadsheets.batchUpdate"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}:batchUpdate"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "BatchUpdateSpreadsheetRequest"
#>   ..$ response  : chr "BatchUpdateSpreadsheetResponse"
#>  $ spreadsheets.values.batchGet               :List of 6
#>   ..$ id        : chr "spreadsheets.values.batchGet"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values:batchGet"
#>   ..$ parameters:List of 18
#>   ..$ request   : chr NA
#>   ..$ response  : chr "BatchGetValuesResponse"
#>  $ spreadsheets.values.clear                  :List of 6
#>   ..$ id        : chr "spreadsheets.values.clear"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values/{range}:clear"
#>   ..$ parameters:List of 15
#>   ..$ request   : chr "ClearValuesRequest"
#>   ..$ response  : chr "ClearValuesResponse"
#>  $ spreadsheets.values.batchClearByDataFilter :List of 6
#>   ..$ id        : chr "spreadsheets.values.batchClearByDataFilter"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values:batchClearByDataFilter"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "BatchClearValuesByDataFilterRequest"
#>   ..$ response  : chr "BatchClearValuesByDataFilterResponse"
#>  $ spreadsheets.values.append                 :List of 6
#>   ..$ id        : chr "spreadsheets.values.append"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values/{range}:append"
#>   ..$ parameters:List of 20
#>   ..$ request   : chr "ValueRange"
#>   ..$ response  : chr "AppendValuesResponse"
#>  $ spreadsheets.values.batchGetByDataFilter   :List of 6
#>   ..$ id        : chr "spreadsheets.values.batchGetByDataFilter"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values:batchGetByDataFilter"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "BatchGetValuesByDataFilterRequest"
#>   ..$ response  : chr "BatchGetValuesByDataFilterResponse"
#>  $ spreadsheets.values.batchClear             :List of 6
#>   ..$ id        : chr "spreadsheets.values.batchClear"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values:batchClear"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "BatchClearValuesRequest"
#>   ..$ response  : chr "BatchClearValuesResponse"
#>  $ spreadsheets.values.get                    :List of 6
#>   ..$ id        : chr "spreadsheets.values.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values/{range}"
#>   ..$ parameters:List of 18
#>   ..$ request   : chr NA
#>   ..$ response  : chr "ValueRange"
#>  $ spreadsheets.values.batchUpdateByDataFilter:List of 6
#>   ..$ id        : chr "spreadsheets.values.batchUpdateByDataFilter"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values:batchUpdateByDataFilter"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "BatchUpdateValuesByDataFilterRequest"
#>   ..$ response  : chr "BatchUpdateValuesByDataFilterResponse"
#>  $ spreadsheets.values.update                 :List of 6
#>   ..$ id        : chr "spreadsheets.values.update"
#>   ..$ method    : chr "PUT"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values/{range}"
#>   ..$ parameters:List of 19
#>   ..$ request   : chr "ValueRange"
#>   ..$ response  : chr "UpdateValuesResponse"
#>  $ spreadsheets.values.batchUpdate            :List of 6
#>   ..$ id        : chr "spreadsheets.values.batchUpdate"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "v4/spreadsheets/{spreadsheetId}/values:batchUpdate"
#>   ..$ parameters:List of 14
#>   ..$ request   : chr "BatchUpdateValuesRequest"
#>   ..$ response  : chr "BatchUpdateValuesResponse"

sheets_endpoints("spreadsheets.values.get")

#> $spreadsheets.values.get
#> $spreadsheets.values.get$id
#> [1] "spreadsheets.values.get"
#> 
#> $spreadsheets.values.get$method
#> [1] "GET"
#> 
#> $spreadsheets.values.get$path
#> [1] "v4/spreadsheets/{spreadsheetId}/values/{range}"
#> 
#> $spreadsheets.values.get$parameters
#> $spreadsheets.values.get$parameters$majorDimension
#> $spreadsheets.values.get$parameters$majorDimension$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$majorDimension$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$majorDimension$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$majorDimension$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$majorDimension$enum
#> [1] "DIMENSION_UNSPECIFIED" "ROWS"                  "COLUMNS"              
#> 
#> $spreadsheets.values.get$parameters$majorDimension$description
#> [1] "The major dimension that results should use.\n\nFor example, if the spreadsheet data is: `A1=1,B1=2,A2=3,B2=4`,\nthen requesting `range=A1:B2,majorDimension=ROWS` will return\n`[[1,2],[3,4]]`,\nwhereas requesting `range=A1:B2,majorDimension=COLUMNS` will return\n`[[1,3],[2,4]]`."
#> 
#> 
#> $spreadsheets.values.get$parameters$spreadsheetId
#> $spreadsheets.values.get$parameters$spreadsheetId$location
#> [1] "path"
#> 
#> $spreadsheets.values.get$parameters$spreadsheetId$required
#> [1] TRUE
#> 
#> $spreadsheets.values.get$parameters$spreadsheetId$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$spreadsheetId$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$spreadsheetId$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$spreadsheetId$description
#> [1] "The ID of the spreadsheet to retrieve data from."
#> 
#> 
#> $spreadsheets.values.get$parameters$range
#> $spreadsheets.values.get$parameters$range$location
#> [1] "path"
#> 
#> $spreadsheets.values.get$parameters$range$required
#> [1] TRUE
#> 
#> $spreadsheets.values.get$parameters$range$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$range$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$range$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$range$description
#> [1] "The A1 notation of the values to retrieve."
#> 
#> 
#> $spreadsheets.values.get$parameters$valueRenderOption
#> $spreadsheets.values.get$parameters$valueRenderOption$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$valueRenderOption$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$valueRenderOption$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$valueRenderOption$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$valueRenderOption$enum
#> [1] "FORMATTED_VALUE"   "UNFORMATTED_VALUE" "FORMULA"          
#> 
#> $spreadsheets.values.get$parameters$valueRenderOption$description
#> [1] "How values should be represented in the output.\nThe default render option is ValueRenderOption.FORMATTED_VALUE."
#> 
#> 
#> $spreadsheets.values.get$parameters$dateTimeRenderOption
#> $spreadsheets.values.get$parameters$dateTimeRenderOption$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$dateTimeRenderOption$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$dateTimeRenderOption$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$dateTimeRenderOption$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$dateTimeRenderOption$enum
#> [1] "SERIAL_NUMBER"    "FORMATTED_STRING"
#> 
#> $spreadsheets.values.get$parameters$dateTimeRenderOption$description
#> [1] "How dates, times, and durations should be represented in the output.\nThis is ignored if value_render_option is\nFORMATTED_VALUE.\nThe default dateTime render option is [DateTimeRenderOption.SERIAL_NUMBER]."
#> 
#> 
#> $spreadsheets.values.get$parameters$upload_protocol
#> $spreadsheets.values.get$parameters$upload_protocol$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$upload_protocol$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$upload_protocol$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$upload_protocol$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$upload_protocol$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$upload_protocol$description
#> [1] "Upload protocol for media (e.g. \"raw\", \"multipart\")."
#> 
#> 
#> $spreadsheets.values.get$parameters$prettyPrint
#> $spreadsheets.values.get$parameters$prettyPrint$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$prettyPrint$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $spreadsheets.values.get$parameters$prettyPrint$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$prettyPrint$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> 
#> $spreadsheets.values.get$parameters$uploadType
#> $spreadsheets.values.get$parameters$uploadType$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$uploadType$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$uploadType$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$uploadType$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$uploadType$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$uploadType$description
#> [1] "Legacy upload protocol for media (e.g. \"media\", \"multipart\")."
#> 
#> 
#> $spreadsheets.values.get$parameters$fields
#> $spreadsheets.values.get$parameters$fields$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$fields$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$fields$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$fields$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$fields$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> 
#> $spreadsheets.values.get$parameters$`$.xgafv`
#> $spreadsheets.values.get$parameters$`$.xgafv`$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$`$.xgafv`$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$`$.xgafv`$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$`$.xgafv`$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$`$.xgafv`$enum
#> [1] "1" "2"
#> 
#> $spreadsheets.values.get$parameters$`$.xgafv`$description
#> [1] "V1 error format."
#> 
#> 
#> $spreadsheets.values.get$parameters$callback
#> $spreadsheets.values.get$parameters$callback$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$callback$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$callback$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$callback$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$callback$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$callback$description
#> [1] "JSONP"
#> 
#> 
#> $spreadsheets.values.get$parameters$alt
#> $spreadsheets.values.get$parameters$alt$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$alt$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$alt$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$alt$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$alt$enum
#> [1] "json"  "media" "proto"
#> 
#> $spreadsheets.values.get$parameters$alt$description
#> [1] "Data format for response."
#> 
#> 
#> $spreadsheets.values.get$parameters$key
#> $spreadsheets.values.get$parameters$key$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$key$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$key$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$key$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$key$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> 
#> $spreadsheets.values.get$parameters$access_token
#> $spreadsheets.values.get$parameters$access_token$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$access_token$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$access_token$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$access_token$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$access_token$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$access_token$description
#> [1] "OAuth access token."
#> 
#> 
#> $spreadsheets.values.get$parameters$quotaUser
#> $spreadsheets.values.get$parameters$quotaUser$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$quotaUser$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$quotaUser$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$quotaUser$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$quotaUser$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$quotaUser$description
#> [1] "Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters."
#> 
#> 
#> $spreadsheets.values.get$parameters$pp
#> $spreadsheets.values.get$parameters$pp$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$pp$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$pp$type
#> [1] "boolean"
#> 
#> $spreadsheets.values.get$parameters$pp$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$pp$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$pp$description
#> [1] "Pretty-print response."
#> 
#> 
#> $spreadsheets.values.get$parameters$oauth_token
#> $spreadsheets.values.get$parameters$oauth_token$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$oauth_token$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$oauth_token$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$oauth_token$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$oauth_token$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> 
#> $spreadsheets.values.get$parameters$bearer_token
#> $spreadsheets.values.get$parameters$bearer_token$location
#> [1] "query"
#> 
#> $spreadsheets.values.get$parameters$bearer_token$required
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$bearer_token$type
#> [1] "string"
#> 
#> $spreadsheets.values.get$parameters$bearer_token$repeated
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$bearer_token$enum
#> [1] NA
#> 
#> $spreadsheets.values.get$parameters$bearer_token$description
#> [1] "OAuth bearer token."
#> 
#> 
#> 
#> $spreadsheets.values.get$request
#> [1] NA
#> 
#> $spreadsheets.values.get$response
#> [1] "ValueRange"
#> 
#> 

sheets_endpoints(4)

#> $spreadsheets.batchUpdate
#> $spreadsheets.batchUpdate$id
#> [1] "spreadsheets.batchUpdate"
#> 
#> $spreadsheets.batchUpdate$method
#> [1] "POST"
#> 
#> $spreadsheets.batchUpdate$path
#> [1] "v4/spreadsheets/{spreadsheetId}:batchUpdate"
#> 
#> $spreadsheets.batchUpdate$parameters
#> $spreadsheets.batchUpdate$parameters$spreadsheetId
#> $spreadsheets.batchUpdate$parameters$spreadsheetId$location
#> [1] "path"
#> 
#> $spreadsheets.batchUpdate$parameters$spreadsheetId$required
#> [1] TRUE
#> 
#> $spreadsheets.batchUpdate$parameters$spreadsheetId$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$spreadsheetId$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$spreadsheetId$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$spreadsheetId$description
#> [1] "The spreadsheet to apply the updates to."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$upload_protocol
#> $spreadsheets.batchUpdate$parameters$upload_protocol$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$upload_protocol$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$upload_protocol$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$upload_protocol$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$upload_protocol$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$upload_protocol$description
#> [1] "Upload protocol for media (e.g. \"raw\", \"multipart\")."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$prettyPrint
#> $spreadsheets.batchUpdate$parameters$prettyPrint$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$prettyPrint$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $spreadsheets.batchUpdate$parameters$prettyPrint$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$prettyPrint$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$uploadType
#> $spreadsheets.batchUpdate$parameters$uploadType$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$uploadType$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$uploadType$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$uploadType$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$uploadType$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$uploadType$description
#> [1] "Legacy upload protocol for media (e.g. \"media\", \"multipart\")."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$fields
#> $spreadsheets.batchUpdate$parameters$fields$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$fields$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$fields$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$fields$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$fields$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`$enum
#> [1] "1" "2"
#> 
#> $spreadsheets.batchUpdate$parameters$`$.xgafv`$description
#> [1] "V1 error format."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$callback
#> $spreadsheets.batchUpdate$parameters$callback$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$callback$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$callback$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$callback$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$callback$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$callback$description
#> [1] "JSONP"
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$alt
#> $spreadsheets.batchUpdate$parameters$alt$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$alt$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$alt$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$alt$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$alt$enum
#> [1] "json"  "media" "proto"
#> 
#> $spreadsheets.batchUpdate$parameters$alt$description
#> [1] "Data format for response."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$key
#> $spreadsheets.batchUpdate$parameters$key$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$key$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$key$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$key$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$key$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$access_token
#> $spreadsheets.batchUpdate$parameters$access_token$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$access_token$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$access_token$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$access_token$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$access_token$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$access_token$description
#> [1] "OAuth access token."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$quotaUser
#> $spreadsheets.batchUpdate$parameters$quotaUser$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$quotaUser$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$quotaUser$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$quotaUser$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$quotaUser$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$quotaUser$description
#> [1] "Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$pp
#> $spreadsheets.batchUpdate$parameters$pp$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$pp$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$pp$type
#> [1] "boolean"
#> 
#> $spreadsheets.batchUpdate$parameters$pp$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$pp$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$pp$description
#> [1] "Pretty-print response."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$oauth_token
#> $spreadsheets.batchUpdate$parameters$oauth_token$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$oauth_token$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$oauth_token$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$oauth_token$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$oauth_token$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> 
#> $spreadsheets.batchUpdate$parameters$bearer_token
#> $spreadsheets.batchUpdate$parameters$bearer_token$location
#> [1] "query"
#> 
#> $spreadsheets.batchUpdate$parameters$bearer_token$required
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$bearer_token$type
#> [1] "string"
#> 
#> $spreadsheets.batchUpdate$parameters$bearer_token$repeated
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$bearer_token$enum
#> [1] NA
#> 
#> $spreadsheets.batchUpdate$parameters$bearer_token$description
#> [1] "OAuth bearer token."
#> 
#> 
#> 
#> $spreadsheets.batchUpdate$request
#> [1] "BatchUpdateSpreadsheetRequest"
#> 
#> $spreadsheets.batchUpdate$response
#> [1] "BatchUpdateSpreadsheetResponse"
#> 
#>