Access IDs of example Sheets
Source: R/sheets_example.R

googlesheets4 ships with static IDs for some world-readable example Sheets for use in examples and documentation. This function make them easy to access by a nickname.

sheets_example(name = NULL)

Arguments

name
Nickname of an example Sheet. If NULL, the examples are listed.

Value

Either a single sheets_id object or a named character vector of all built-in examples.

Examples

sheets_example()

#>                                      gapminder 
#> "1BzfL0kZUz1TsI5zxJF1WNF01IxvC67FbOJUiiGMZ_mQ" 
#>                                       mini-gap 
#> "1BMtx1V2pk2KG2HGANvvBOaZM4Jx1DUdRrFdEx-OJIGY" 
#>                                             ff 
#> "132Ij_8ggTKVLnLqCOM3ima6mV9F8rmY7HEcR-5hjWoQ" 
#>                                   design-dates 
#> "1xTUxWGcFLtDIHoYJ1WsjQuLmpUtBf--8Bcu5lQ302SU" 
#>                                         deaths 
#> "1ESTf_tH08qzWwFYRC1NVWJjswtLdZn9EGw5e3Z5wMzA" 

sheets_example("gapminder")

#> [1] "1BzfL0kZUz1TsI5zxJF1WNF01IxvC67FbOJUiiGMZ_mQ"
#> attr(,"class")
#> [1] "sheets_id" "drive_id"