Find Google Sheets

Finds your Google Sheets. This is a very thin wrapper around googledrive::drive_find(), that specifies you want to list Drive files where type = "spreadsheet". Therefore, note that this will require auth for googledrive! See the article Using googlesheets4 with googledrive if you want to coordinate auth between googlesheets4 and googledrive.

sheets_find(...)

Arguments

...
Arguments (other than type, which is hard-wired as type = "spreadsheet") that are passed along to googledrive::drive_find().

Value

An object of class dribble, a tibble with one row per item.

Examples

if (sheets_has_token()) {
  # see all your Sheets
  sheets_find()

  # see 5 Sheets, prioritized by creation time
  x <- sheets_find(order_by = "createdTime desc", n_max = 5)
  x

  # hoist the creation date, using other packages in the tidyverse
  # x %>%
  #   tidyr::hoist(drive_resource, created_on = "createdTime") %>%
  #   dplyr::mutate(created_on = as.Date(created_on))
}

#> # A tibble: 5 x 3
#>   name               id                                         drive_resource  
#> * <chr>              <chr>                                      <list>          
#> 1 uncheery-nudibran… 193Apymy0IFFvveYO7L8pwSxQrp4ntZX4LqhO-dJP… <named list [34…
#> 2 befuddling-skink   12CJhr6nixPhWQi-LWKKQtRK1lAmbH3l6QWWcdNpo… <named list [34…
#> 3 cubical-barbet     1UVqMf9Q-mcIV9p7hRFdkgzBRjNgS_IQYAo9t23pJ… <named list [34…
#> 4 holy-chamois       1wCV7ryNXOlZyNw-HDZnPOqmTLuwVp1403It9h_Sk… <named list [34…
#> 5 arboresque-bass    1V3e5nhgFZja4gxobe1oTykXa_tq6SQbtNYyI0iWP… <named list [34…