Get data about (work)sheets

Reveals full metadata or just the names for the (work)sheets inside a (spread)Sheet.

sheets_sheet_data(ss)

sheets_sheet_names(ss)

Arguments

ss
Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get(), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().

Value

sheets_sheet_data(): A tibble with one row per (work)sheet.

sheets_sheet_names(): A character vector of (work)sheet names.

See also

Examples

if (sheets_has_token()) {
  ss <- sheets_example("gapminder")
  sheets_sheet_data(ss)
  sheets_sheet_names(ss)
}

#> [1] "Africa"   "Americas" "Asia"     "Europe"   "Oceania"