Delete one or more (work)sheets

Deletes one or more (work)sheets from a (spread)Sheet.

sheets_sheet_delete(ss, sheet)

Arguments

ss
Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by sheets_get(), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().

sheet
Sheet to delete, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. You can pass a vector to delete multiple sheets at once or even a list, if you need to mix names and positions.

Value

The input ss, as an instance of sheets_id

See also

Examples

if (sheets_has_token()) {
  ss <- sheets_create("delete-sheets-from-me")
  sheets_sheet_add(ss, c("alpha", "beta", "gamma", "delta"))

  # get an overview of the sheets
  sheets_sheet_data(ss)

  # delete sheets
  sheets_sheet_delete(ss, 1)
  sheets_sheet_delete(ss, "gamma")
  sheets_sheet_delete(ss, list("alpha", 2))

  # get an overview of the sheets
  sheets_sheet_data(ss)

  # cleanup
  sheets_find("delete-sheets-from-me") %>% googledrive::drive_rm()
}

#> Adding these sheets:
#> * alpha
#> * beta
#> * gamma
#> * delta

#> Deleting these sheet(s) from 'delete-sheets-from-me':
#> * Sheet1

#> Deleting these sheet(s) from 'delete-sheets-from-me':
#> * gamma

#> Deleting these sheet(s) from 'delete-sheets-from-me':
#> * alpha
#> * beta

#> Files deleted:
#>   * delete-sheets-from-me: 1PsYOAxTKOMCe1_c9toUEIg_WtuFeQDjpNXdQ4edLlrg