Get info on current user

Reveals the email address of the user associated with the current token. If no token has been loaded yet, this function does not initiate auth.

sheets_user()

Value

An email address or, if no token has been loaded, NULL.

See also

Examples

sheets_user()

#> [1] "googlesheets4-testing@gargle-169921.iam.gserviceaccount.com"