Coerce to a sheets_id object

Converts various representations of a Google Sheet into a sheets_id object. Anticipated inputs:

Spreadsheet id, "a string containing letters, numbers, and some special characters", typically 44 characters long, in our experience. Example: 1qpyC0XzvTcKT6EISywvqESX3A0MwQoFDE8p-Bll4hps.

A URL, from which we can excavate a spreadsheet or file id. Example: "https://docs.google.com/spreadsheets/d/1BzfL0kZUz1TsI5zxJF1WNF01IxvC67FbOJUiiGMZ_mQ/edit#gid=1150108545".

A one-row dribble, a "Drive tibble" used by the googledrive package. In general, a dribble can represent several files, one row per file. Since googlesheets4 is not vectorized over spreadsheets, we are only prepared to accept a one-row dribble.

googledrive::drive_get("YOUR_SHEET_NAME") is a great way to look up a Sheet via its name.

gs4_find("YOUR_SHEET_NAME") is another good way to get your hands on a Sheet.

Spreadsheet meta data, as returned by, e.g., gs4_get(). Literally, this is an object of class googlesheets4_spreadsheet.

This is a generic function.

as_sheets_id(x, ...)

Arguments

x
Something that uniquely identifies a Google Sheet: a sheets_id, a drive_id, a URL, a one-row dribble, or a googlesheets4_spreadsheet.

...
Other arguments passed down to methods. (Not used.)

Examples

as_sheets_id("abc")

#> Spreadsheet name: <unknown>
#>               ID: abc
#>           Locale: <unknown>
#>        Time zone: <unknown>
#>      # of sheets: <unknown>
#> 
#> Unable to get metadata for this Sheet. Error details:
#> Client error: (404) NOT_FOUND
#> * A specified resource is not found, or the request is rejected by undisclosed reasons, such as whitelisting.
#> * Requested entity was not found.