Specify cells for reading

The range argument in read_sheet() or sheets_cells() is used to limit the read to a specific rectangle of cells. The Sheets v4 API only accepts ranges in A1 notation, but googlesheets4 accepts and converts a few alternative specifications provided by the functions in the cellranger package. Of course, you can always provide A1-style ranges directly to functions like read_sheet() or sheets_cells(). Why would you use the cellranger helpers? Some ranges are practically impossible to express in A1 notation, specifically when you want to describe rectangles with some bounds that are specified and others determined by the data.

Examples

if (sheets_has_token()) {
  ss <- sheets_example("mini-gap")

  # Specify only the rows or only the columns
  read_sheet(ss, range = cell_rows(1:3))
  read_sheet(ss, range = cell_cols("C:D"))
  read_sheet(ss, range = cell_cols(1))

  # Specify upper or lower bound on row or column
  read_sheet(ss, range = cell_rows(c(NA, 4)))
  read_sheet(ss, range = cell_cols(c(NA, "D")))
  read_sheet(ss, range = cell_rows(c(3, NA)))
  read_sheet(ss, range = cell_cols(c(2, NA)))
  read_sheet(ss, range = cell_cols(c("C", NA)))

  # Specify a partially open rectangle
  read_sheet(ss, range = cell_limits(c(2, 3), c(NA, NA)), col_names = FALSE)
  read_sheet(ss, range = cell_limits(c(1, 2), c(NA, 4)))
}

#> Reading from 'mini-gap'

#> Range "1:3"

#> Reading from 'mini-gap'

#> Range "C:D"

#> Reading from 'mini-gap'

#> Range "A:A"

#> Reading from 'mini-gap'

#> Range "1:4"

#> Reading from 'mini-gap'

#> Range "A:D"

#> Reading from 'mini-gap'

#> Range "3:5000000"

#> Reading from 'mini-gap'

#> Range "B:ZZZ"

#> Reading from 'mini-gap'

#> Range "C:ZZZ"

#> Reading from 'mini-gap'

#> Range "C2:ZZZ"

#> New names:
#> * `` -> ...1
#> * `` -> ...2
#> * `` -> ...3
#> * `` -> ...4

#> Reading from 'mini-gap'

#> Range "B1:D"

#> # A tibble: 5 x 3
#>   continent  year lifeExp
#>   <chr>     <dbl>   <dbl>
#> 1 Africa     1952    43.1
#> 2 Africa     1952    30.0
#> 3 Africa     1952    38.2
#> 4 Africa     1952    47.6
#> 5 Africa     1952    32.0