googlesheets4-package {googlesheets4}	R Documentation

googlesheets4: Access Google Sheets using the Sheets API V4

Description

Interact with Google Sheets through the Sheets API v4 <https://developers.google.com/sheets/api>. "API" is an acronym for "application programming interface"; the Sheets API allows users to interact with Google Sheets programmatically, instead of via a web browser. The "v4" refers to the fact that the Sheets API is currently at version 4. This package can read and write both the metadata and the cell data in a Sheet.

Author(s)

Maintainer: Jennifer Bryan jenny@rstudio.com (ORCID)

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://googlesheets4.tidyverse.org

https://github.com/tidyverse/googlesheets4

Report bugs at https://github.com/tidyverse/googlesheets4/issues

[Package googlesheets4 version 1.0.0 Index]