gs4_browse {googlesheets4}	R Documentation

Visit a Sheet in a web browser

Description

Visits a Google Sheet in your default browser, if session is interactive.

Usage

gs4_browse(ss)

Arguments

ss
Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

Value

The Sheet's browser URL, invisibly.

Examples

gs4_example("mini-gap") %>% gs4_browse()

[Package googlesheets4 version 1.0.0 Index]