gs4_create {googlesheets4}	R Documentation

Create a new Sheet

Description

Creates an entirely new (spread)Sheet (or, in Excel-speak, workbook). Optionally, you can also provide names and/or data for the initial set of (work)sheets. Any initial data provided via sheets is styled as a table, as described in sheet_write().

Usage

gs4_create(name = gs4_random(), ..., sheets = NULL)

Arguments

name
The name of the new spreadsheet.

...
Optional spreadsheet properties that can be set through this API endpoint, such as locale and time zone.

sheets
Optional input for initializing (work)sheets. If unspecified, the Sheets API automatically creates an empty "Sheet1". You can provide a vector of sheet names, a data frame, or a (possibly named) list of data frames. See the examples.

Value

The input ss, as an instance of sheets_id

See Also

Wraps the spreadsheets.create endpoint:

https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/create

There is an article on writing Sheets:

https://googlesheets4.tidyverse.org/articles/articles/write-sheets.html

Other write functions: gs4_formula(), range_delete(), range_flood(), range_write(), sheet_append(), sheet_write()

Examples

if (gs4_has_token()) {
  gs4_create("gs4-create-demo-1")

  gs4_create("gs4-create-demo-2", locale = "en_CA")

  gs4_create(
    "gs4-create-demo-3",
    locale = "fr_FR",
    timeZone = "Europe/Paris"
  )

  gs4_create(
    "gs4-create-demo-4",
    sheets = c("alpha", "beta")
  )

  my_data <- data.frame(x = 1)
  gs4_create(
    "gs4-create-demo-5",
    sheets = my_data
  )

  gs4_create(
    "gs4-create-demo-6",
    sheets = list(chickwts = head(chickwts), mtcars = head(mtcars))
  )

  # clean up
  gs4_find("gs4-create-demo") %>%
    googledrive::drive_trash()
}

[Package googlesheets4 version 1.0.0 Index]