Create a new Sheet

Creates an entirely new (spread)Sheet (or, in Excel-speak, workbook). Optionally, you can also provide names and/or data for the initial set of (work)sheets. Any initial data provided via sheets is styled as a table, as described in sheet_write().

Usage

gs4_create(name = gs4_random(), ..., sheets = NULL)

Arguments

name

The name of the new spreadsheet.

...

Optional spreadsheet properties that can be set through this API endpoint, such as locale and time zone.

sheets

Optional input for initializing (work)sheets. If unspecified, the Sheets API automatically creates an empty "Sheet1". You can provide a vector of sheet names, a data frame, or a (possibly named) list of data frames. See the examples.

Value

The input ss, as an instance of sheets_id

See also

Examples

gs4_create("gs4-create-demo-1")
#> ✔ Creating new Sheet: gs4-create-demo-1.

gs4_create("gs4-create-demo-2", locale = "en_CA")
#> ✔ Creating new Sheet: gs4-create-demo-2.

gs4_create(
  "gs4-create-demo-3",
  locale = "fr_FR",
  timeZone = "Europe/Paris"
)
#> ✔ Creating new Sheet: gs4-create-demo-3.

gs4_create(
  "gs4-create-demo-4",
  sheets = c("alpha", "beta")
)
#> ✔ Creating new Sheet: gs4-create-demo-4.

my_data <- data.frame(x = 1)
gs4_create(
  "gs4-create-demo-5",
  sheets = my_data
)
#> ✔ Creating new Sheet: gs4-create-demo-5.

gs4_create(
  "gs4-create-demo-6",
  sheets = list(chickwts = head(chickwts), mtcars = head(mtcars))
)
#> ✔ Creating new Sheet: gs4-create-demo-6.

# Clean up
gs4_find("gs4-create-demo") %>%
  googledrive::drive_trash()
#> Files trashed:
#> • gs4-create-demo-6 <id: 1ZAMtYbi5sDjQMChplU6XfC1T7IKox5ry0Pi7usiyCuw>
#> • gs4-create-demo-5 <id: 1x2uV_yoCON9ebVT8xL8CrbI7il9sjMJ9IcGbN15_yxc>
#> • gs4-create-demo-4 <id: 12h1Fig3u5zLr-kpWZFnhz_cdblrEfJDAEP_I9NYYOdQ>
#> • gs4-create-demo-3 <id: 1NWL7Ec8oCVrdJJe3FNcUlQxDvMcxKMvfE3SIP92vIj8>
#> • gs4-create-demo-2 <id: 1co4zZLGzdQQs0TMRaWDehehf7BUocTJXr0wsfk3O8Hc>
#> • gs4-create-demo-1 <id: 1xddCKBlrKGfNeAn0Mwyj-yZQRV8XhB0qBXMGf_Tqzbw>