gs4_examples {googlesheets4}	R Documentation

Example Sheets

Description

googlesheets4 makes a variety of world-readable example Sheets available for use in documentation and reprexes. These functions help you access the example Sheets. See vignette("example-sheets", package = "googlesheets4") for more.

Usage

gs4_examples(matches)

gs4_example(matches)

Arguments

matches
A regular expression that matches the name of the desired example Sheet(s). matches is optional for the plural gs4_examples() and, if provided, it can match multiple Sheets. The singular gs4_example() requires matches and it must match exactly one Sheet.

Value

gs4_example(): a sheets_id

gs4_examples(): a named vector of all built-in examples, with class drive_id

Examples

gs4_examples()
gs4_examples("gap")

gs4_example("gapminder")
gs4_example("deaths")

[Package googlesheets4 version 1.1.1 Index]