gs4_fodder {googlesheets4}	R Documentation

Create useful spreadsheet filler

Description

Creates a data frame that is useful for filling a spreadsheet, when you just need a sheet to experiment with. The data frame has n rows and m columns with these properties:

Column names match what Sheets displays: "A", "B", "C", and so on.

Inner cell values reflect the coordinates where each value will land in the sheet, in A1-notation. So the first row is "B2", "C2", and so on. Note that this n-row data frame will occupy n + 1 rows in the sheet, because the column names occupy the first row.

Usage

gs4_fodder(n = 10, m = n)

Arguments

n
Number of rows.

m
Number of columns.

Value

A data frame of character vectors.

Examples

gs4_fodder()
gs4_fodder(5, 3)

[Package googlesheets4 version 1.0.0 Index]