sheet_add {googlesheets4}	R Documentation

Add one or more (work)sheets

Description

Adds one or more (work)sheets to an existing (spread)Sheet. Note that sheet names must be unique.

Usage

sheet_add(ss, sheet = NULL, ..., .before = NULL, .after = NULL)

Arguments

ss
Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

sheet
One or more new sheet names. If unspecified, one new sheet is added and Sheets autogenerates a name of the form "SheetN".

...
Optional parameters to specify additional properties, common to all of the new sheet(s). Not relevant to most users. Specify fields of the SheetProperties schema in name = value form.

.before, .after
Optional specification of where to put the new sheet(s). Specify, at most, one of .before and .after. Refer to an existing sheet by name (via a string) or by position (via a number). If unspecified, Sheets puts the new sheet(s) at the end.

Value

The input ss, as an instance of sheets_id

See Also

Makes a batch of AddSheetRequests (one per sheet):

https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#addsheetrequest

Other worksheet functions: sheet_append(), sheet_copy(), sheet_delete(), sheet_properties(), sheet_relocate(), sheet_rename(), sheet_resize(), sheet_write()

Examples

if (gs4_has_token()) {
  ss <- gs4_create("add-sheets-to-me")

  # the only required argument is the target spreadsheet
  ss %>% sheet_add()

  # but you CAN specify sheet name and/or position
  ss %>% sheet_add("apple", .after = 1)
  ss %>% sheet_add("banana", .after = "apple")

  # add multiple sheets at once
  ss %>% sheet_add(c("coconut", "dragonfruit"))

  # keeners can even specify additional sheet properties
  ss %>%
    sheet_add(
      sheet = "eggplant",
      .before = 1,
      gridProperties = list(
        rowCount = 3, columnCount = 6, frozenRowCount = 1
      )
    )

  # get an overview of the sheets
  sheet_properties(ss)

  # clean up
  gs4_find("add-sheets-to-me") %>%
    googledrive::drive_trash()
}

[Package googlesheets4 version 1.0.0 Index]