Delete one or more (work)sheets

Deletes one or more (work)sheets from a (spread)Sheet.

sheet_delete(ss, sheet)

Arguments

ss
Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by gs4_get()), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().

sheet
Sheet to delete, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. You can pass a vector to delete multiple sheets at once or even a list, if you need to mix names and positions.

Value

The input ss, as an instance of sheets_id

See also

Examples

if (gs4_has_token()) {
  ss <- gs4_create("delete-sheets-from-me")
  sheet_add(ss, c("alpha", "beta", "gamma", "delta"))

  # get an overview of the sheets
  sheet_properties(ss)

  # delete sheets
  sheet_delete(ss, 1)
  sheet_delete(ss, "gamma")
  sheet_delete(ss, list("alpha", 2))

  # get an overview of the sheets
  sheet_properties(ss)

  # clean up
  gs4_find("delete-sheets-from-me") %>%
    googledrive::drive_trash()
}

#> ✔ Creating new Sheet: delete-sheets-from-me

#> ✔ Adding 4 sheets to delete-sheets-from-me

#> ✔ New sheets: alpha, beta, gamma, and delta

#> ✔ Deleting 1 sheet from delete-sheets-from-me: Sheet1

#> ✔ Deleting 1 sheet from delete-sheets-from-me: gamma

#> ✔ Deleting 2 sheets from delete-sheets-from-me: alpha and beta

#> File trashed:

#> • delete-sheets-from-me <id: 1ShdxTlQrOzh9YVNtvP75l1nVuAxOECngnLVtKIVh1ik>