sheet_properties {googlesheets4}	R Documentation

Get data about (work)sheets

Description

Reveals full metadata or just the names for the (work)sheets inside a (spread)Sheet.

Usage

sheet_properties(ss)

sheet_names(ss)

Arguments

ss
Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

Value

sheet_properties(): A tibble with one row per (work)sheet.

sheet_names(): A character vector of (work)sheet names.

See Also

Other worksheet functions: sheet_add(), sheet_append(), sheet_copy(), sheet_delete(), sheet_relocate(), sheet_rename(), sheet_resize(), sheet_write()

Examples

ss <- gs4_example("gapminder")
sheet_properties(ss)
sheet_names(ss)

[Package googlesheets4 version 1.1.1 Index]