sheet_rename {googlesheets4}	R Documentation

Rename a (work)sheet

Description

Changes the name of a (work)sheet.

Usage

sheet_rename(ss, sheet = NULL, new_name)

Arguments

ss
Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

sheet
Sheet to rename, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Defaults to the first visible sheet.

new_name
New name of the sheet, as a string. This is required.

Value

The input ss, as an instance of sheets_id

See Also

Makes an UpdateSheetPropertiesRequest:

https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#UpdateSheetPropertiesRequest

Other worksheet functions: sheet_add(), sheet_append(), sheet_copy(), sheet_delete(), sheet_properties(), sheet_relocate(), sheet_resize(), sheet_write()

Examples

if (gs4_has_token()) {
  ss <- gs4_create(
    "sheet-rename-demo",
    sheets = list(cars = head(cars), chickwts = head(chickwts))
  )
  sheet_names(ss)

  ss %>%
    sheet_rename(1, new_name = "automobiles") %>%
    sheet_rename("chickwts", new_name = "poultry")

  # clean up
  gs4_find("sheet-rename-demo") %>%
    googledrive::drive_trash()
}

[Package googlesheets4 version 1.0.0 Index]