Create a new Sheet

Creates an entirely new Sheet (spreadsheet or workbook). Offers some control over the initial set of sheets (worksheets or tabs). CAUTION: this function is still being developed and, for example, currently sends all data as character.

sheets_create(name, ..., sheets = NULL)

Arguments

name
The name of the spreadsheet.

...
Optional spreadsheet properties that can be set through this API endpoint, such as locale and time zone.

sheets
Optional named list of data frames. One sheet is created for each data frame.

Value

sheets_get(): A list with S3 class sheets_Spreadsheet, for printing purposes.

sheets_sheets(): A character vector.

See also

Examples

if (sheets_has_token()) {
  sheets_create("sheets-create-demo-1")

  sheets_create("sheets-create-demo-2", locale = "en_CA")

  sheets_create(
    "sheets-create-demo-3",
    locale = "fr_FR",
    timeZone = "Europe/Paris"
  )

  sheets_create(
    "sheets-create-demo-4",
    sheets = list(iris = head(iris), mtcars = head(mtcars))
  )

  # clean up
  sheets_find("sheets-create-demo") %>% googledrive::drive_trash()
}

#> Files trashed:
#>   * sheets-create-demo-4: 1i7WUNsepoN8_nm5qTDZrQqgAnZtIAYr_4G6H5XrdC9g
#>   * sheets-create-demo-3: 1nxtpfsOE9rxpjyys_f3KEFodxrBokVTeeKuSx6jq2qo
#>   * sheets-create-demo-2: 1irsBHJE1IQ3NGUqp_4QwzCAwGb2LeI8lgx5g6n9lYCw
#>   * sheets-create-demo-1: 14L-AAsV9xpj67eCh7tmq5WBqLFHgQIOIb3PQpe71U0g