Access IDs of example Sheets

googlesheets4 ships with static IDs for some world-readable example Sheets for use in examples and documentation. These functions make them easy to access by a nickname.

sheets_example(name = names(sheets_examples()))

sheets_examples()

Arguments

name
Nickname of an example Sheet.

Value

sheets_example(): a single sheets_id object

sheets_examples(): a named character vector of all built-in examples

Examples

sheets_examples()

#>                                      gapminder 
#> "1U6Cf_qEOhiR9AZqTqS3mbMF3zt2db48ZP5v3rkrAEJY" 
#>                                       mini-gap 
#> "1k94ZVVl6sdj0AXfK9MQOuQ4rOhd1PULqpAu2_kr9MAU" 
#>                           formulas-and-formats 
#> "1wPLrWOxxEjp3T1nv2YBxn63FX70Mz5W5Tm4tGc-lRms" 
#>                      cell-contents-and-formats 
#> "1peJXEeAp5Qt3ENoTvkhvenQ36N3kLyq6sq9Dh2ufQ6E" 
#>                                         deaths 
#> "1tuYKzSbLukDLe5ymf_ZKdQA8SfOyeMM7rmf6D6NJpxg" 
#>                                  chicken-sheet 
#> "1ct9t1Efv8pAGN9YO5gC2QfRq2wT4XjNoTMXpVeUghJU" 
#> attr(,"class")
#> [1] "sheets_id" "drive_id" 

sheets_example("gapminder")

#> [1] "1U6Cf_qEOhiR9AZqTqS3mbMF3zt2db48ZP5v3rkrAEJY"
#> attr(,"class")
#> [1] "sheets_id" "drive_id"