(Over)write new data into an existing sheet

Updates an existing (work)sheet in an existing (spread)Sheet with a data frame. All pre-existing values, formats, and dimensions of the targeted sheet are cleared and it gets its new values and dimensions from data. Special formatting is applied to the header row, which holds column names, and the first skip + 1 rows are frozen (so, up to and including the header row). NOTE: this function is very alpha and currently writes everything as character.

write_sheet(data, ss, sheet = NULL, skip = 0, na = "")

sheets_write(data, ss, sheet = NULL, skip = 0, na = "")

Arguments

data
A data frame.

ss
Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().

sheet
Sheet to write into, as in "worksheet" or "tab". Either a string (the name of a sheet), or an integer (the position of the sheet). If unspecified, defaults to the first sheet.

skip
Number of rows to leave empty before starting to write.

na
Not implemented yet.

Value

Updated metadata for the (spread)Sheet ss, as an instance of S3 class googlesheets4_spreadsheet.

Examples

if (sheets_has_token()) {
  # create a Sheet with some initial, placeholder data
  ss <- sheets_create(
    "sheets-write-demo",
    sheets = list(alpha = data.frame(x = 1), omega = data.frame(x = 1))
  )

  df <- data.frame(
    x = 1:3,
    y = letters[1:3],
    stringsAsFactors = FALSE
  )

  # write df into the first sheet
  (ss <- sheets_write(data = df, ss = ss))

  # write mtcars into the sheet named 'omega'
  (ss <- sheets_write(data = mtcars, ss = ss, sheet = "omega"))

  # view your magnificent creation in the browser
  # sheets_browse(ss)

  # clean up
  sheets_find("sheets-write-demo") %>% googledrive::drive_rm()
}

#> Writing to 'sheets-write-demo'

#> Writing to sheet "alpha"

#> Writing to 'sheets-write-demo'

#> Writing to sheet "omega"

#> Files deleted:
#>   * sheets-write-demo: 1Z6NNjPPVNaUScEsC1eHf5-uuspDKceCGp1Z2WEGfuvo